requirement.txt
h5py
numpy
torch==1.4.0
torchvision==0.2.0
pyyaml
enum34
future
scipy==1.4.1
opencv_contrib_python==3.4.2.17
matplotlib==3.0.2
transforms3d==0.3.1

lmdb==0.94
setuptools==41.0.0
cffi==1.11.5
easydict==1.7
plyfile==0.6
glumpy==1.0.6
pillow==8.2.0
